\name{NPCStochOrd-package}
\alias{NPCStochOrd-package}

\title{NPCStochOrd: NonParametric Combination (NPC) methodology for Stochastic Ordering problems}

\description{
NPCStochOrd provides NonParametric Combination (NPC) methodology-based tests to address \emph{2}- and \emph{C}-sample stochastic ordering problems in the presence of univariate and multivariate outcomes.}

\author{
Riccardo Ceccato, Luigi Salmaso
}

\references{
Pesarin, F., & Salmaso, L. (2010). \emph{Permutation tests for complex data: theory, applications and software}. John Wiley & Sons.

Arboretti, R., Ceccato, R., & Salmaso, L. (2021). Permutation testing for goodness-of-fit and stochastic ordering with multivariate mixed variables. \emph{Journal of Statistical Computation and Simulation}, 91(5), 876-896.
}

\seealso{
\code{\link{test.stat.C.samples}, \link{test.stat.2.samples}}
}
